refactor: extract core logic into pkg/slirpnetstack library package
Move all core logic from root package main into
pkg/slirpnetstack as a reusable library. Root main.go is now a
thin CLI entry point calling slirpnetstack.Main(). Module path
changed from github.com/cloudflare/slirpnetstack to
github.com/KusakabeShi/slirpnetstack.
