+++ pkg/slirpnetstack/fwd.go
@@ -1,4 +1,4 @@
-package main
+package slirpnetstack
 
 import (
 	"fmt"
