+++ pkg/slirpnetstack/gateway_test.go
@@ -1,4 +1,4 @@
-package main
+package slirpnetstack
 
 import (
 	"testing"
