+++ pkg/slirpnetstack/stack.go
@@ -1,4 +1,4 @@
-package main
+package slirpnetstack
 
 import (
 	"errors"
@@ -143,7 +143,7 @@ func createLinkEP(s *stack.Stack, tunFd int, tapMode bool, macAddress net.Hardwa
 
 func createNIC(s *stack.Stack, nic tcpip.NICID, linkEP stack.LinkEndpoint) error {
 	if err := s.CreateNIC(nic, linkEP); err != nil {
-		fmt.Fprintf(os.Stderr, "[!] CreateNIC(%s) = %s\n", ifName, err)
+		fmt.Fprintf(os.Stderr, "[!] CreateNIC(%d) = %s\n", nic, err)
 		return fmt.Errorf("%s", err)
 	}
